feat: added protected implementations

+++ libs/models/order-book.go
@@ -1,10 +1,10 @@
 package models
 
 type OrderBook struct {
-	OrderId int  `json:"order_id"`
-	BookId  int  `json:"book_id"`
-	Amount  uint `json:"amount"`
-	Price   uint `json:"price"`
+	OrderId int     `json:"order_id"`
+	BookId  int     `json:"book_id"`
+	Amount  uint    `json:"amount"`
+	Price   float32 `json:"price"`
 }
 
 func (OrderBook) TableName() string {
